refactor: Make LabelType non internal
