+++ packages/core/src/Components/Label/Internal/Lib/UseLabelType.ts
@@ -1,4 +1,4 @@
-import type { LabelType }                     from '@/Components/Label/Internal';
+import type { LabelType }                     from '@/Components/Label';
 import type { LabellableTypeElementName }     from '@/Shared/Types/Variants';
 import type { ComputedRef, MaybeRefOrGetter } from 'vue';
 import { useElementClass }                    from '@/Shared/Internal/Lib/UseElementClass';
