--- packages/core/src/Components/Label/Internal/Types/index.ts
@@ -1,1 +1,0 @@
-export * from '@/Components/Label/Internal/Types/Label';
