+++ packages/core/src/Components/Label/Internal/index.ts
@@ -1,3 +1,2 @@
 export * from '@/Components/Label/Internal/Lib';
-export * from '@/Components/Label/Internal/Types';
 export * from '@/Components/Label/Internal/UI';
