+++ packages/core/src/Components/Label/Types/Label.ts
