+++ packages/core/src/Components/Label/index.ts
@@ -1,0 +1,1 @@
+export * from '@/Components/Label/Types';
