+++ packages/core/src/Components/Select/UI/FoSelect.vue
@@ -39,7 +39,7 @@
 </template>
 
 <script setup lang="ts" generic="T extends string | number, K extends SelectOption<T>">
-import type { LabelType }               from '@/Components/Label/Internal';
+import type { LabelType }               from '@/Components/Label';
 import type { SelectOption }            from '@/Components/Select';
 import type { ElementName, Size }       from '@/Shared/Types/Variants';
 import { FoFilledFocused }              from '@/Components/Focus/Internal';
